+++ packages/wrapper/src/wrapper.ts
@@ -12,7 +12,7 @@ import { Logger } from 'monaco-languageclient/tools';
 import { checkServiceConsistency, configureServices } from './vscode/services.js';
 import { EditorAppExtended } from './editorAppExtended.js';
 import { EditorAppClassic } from './editorAppClassic.js';
-import { CodeResources, ModelRefs } from './editorAppBase.js';
+import { CodeResources, ModelRefs, TextModels } from './editorAppBase.js';
 import { LanguageClientWrapper } from './languageClientWrapper.js';
 import { WorkerConfigDirect, WorkerConfigOptions } from './commonTypes.js';
 import { UserConfig } from './userConfig.js';
@@ -147,8 +147,8 @@ export class MonacoEditorLanguageClientWrapper {
         return this.languageClientWrapper?.getLanguageClient();
     }
 
-    getTextModel(original?: boolean): monaco.editor.ITextModel | undefined {
-        return this.editorApp?.getTextModel(original);
+    getTextModels(): TextModels | undefined {
+        return this.editorApp?.getTextModels();
     }
 
     getModelRefs(): ModelRefs | undefined {