+++ CHANGELOG.md
@@ -7,3 +7,11 @@ All notable changes for `monaco-languageclient` are found here:
 All notable changes for `vscode-ws-jsonrpc` are found here:
 
 [vscode-ws-jsonrpc CHANGELOG](./packages/vscode-ws-jsonrpc/CHANGELOG.md)
+
+All notable changes for `monaco-editor-wrapper` are found here:
+
+[monaco-editor-wrapper CHANGELOG](./packages/wrapper/CHANGELOG.md)
+
+All notable changes for `@typefox/monaco-editor-react` are found here:
+
+[@typefox/monaco-editor-react CHANGELOG](./packages/wrapper-react/CHANGELOG.md)